Review fixes and small refactoring
