+++ duckscript_sdk/src/sdk/std/fs/zip/unpack/mod.rs
@@ -32,7 +32,8 @@ impl Command for CommandImpl {
     fn run(&self, arguments: Vec<String>) -> CommandResult {
         if arguments.len() < 2 {
             return CommandResult::Error(
-                "Paths to the ZIP file and/or target directory are not provided.".to_string());
+                "Paths to the ZIP file and/or target directory are not provided.".to_string(),
+            );
         }
 
         let zipfile = Path::new(&arguments[0]);
@@ -40,36 +41,33 @@ impl Command for CommandImpl {
 
         match std::fs::create_dir_all(target_dir) {
             Ok(_) => (),
-            Err(err) =>
-                return CommandResult::Error(format!("Couldn't create target directory: {}", err)),
+            Err(err) => {
+                return CommandResult::Error(format!("Couldn't create target directory: {}", err))
+            }
         };
 
-        let zip_file = match OpenOptions::new()
-            .read(true)
-            .open(zipfile)
-        {
+        let zip_file = match OpenOptions::new().read(true).open(zipfile) {
             Ok(file) => file,
-            Err(err) =>
-                return CommandResult::Error(format!("Couldn't open ZIP file: {}", err)),
+            Err(err) => return CommandResult::Error(format!("Couldn't open ZIP file: {}", err)),
         };
         let mut zip = match ZipArchive::new(zip_file) {
             Ok(archive) => archive,
-            Err(err) =>
-                return CommandResult::Error(format!("Couldn't open ZIP file: {}", err)),
+            Err(err) => return CommandResult::Error(format!("Couldn't open ZIP file: {}", err)),
         };
 
         for file in zip.file_names() {
             let file_path = target_dir.join(file);
             if file_path.exists() && file_path.is_file() {
-                return CommandResult::Error(
-                    format!("File already exists: {}", file_path.to_str().unwrap()));
+                return CommandResult::Error(format!(
+                    "File already exists: {}",
+                    file_path.to_str().unwrap()
+                ));
             }
         }
 
         match zip.extract(target_dir) {
             Ok(_) => (),
-            Err(err) =>
-                return CommandResult::Error(format!("Couldn't unpack ZIP file: {}", err)),
+            Err(err) => return CommandResult::Error(format!("Couldn't unpack ZIP file: {}", err)),
         };
 
         CommandResult::Continue(Some("true".to_string()))
