+++ test/std/fs/zip/unpack_test.ds
@@ -1,4 +1,4 @@
-base = set ./target/_duckscript_test/zip/unpack
+base = set ./target/_duckscript_test/fs/zip/unpack
 
 fn randfile
     length = random_range 10 50
